requirements.txt
pytest
pylint
radon
requests
python-dotenv
